Refactor into standalone executable
Making tags generation part of the build pipeline gives us the benefit
of accessing packages, but is not very useful or ergonomic as users
should not be expected to modify the build.zig file for each Zig project
they want to generate tags for.

So for now revert back to using a standalone executable. Packages can be
indexed by specifying their entrypoints on the command line.