Use search string as tags command
This is a bit wasteful, as we already know the *exact* byte offset of
the token from the AST, so converting that into searching for a regex
adds a lot of unnecessary work. However, Vim's tags implementation is
optimized for regex based searches, so using search patterns makes a lot
of other things inside Vim work a bit smoother.