Remove assumption that an arena allocator was always used
Originally I omitted almost all of the resource/memory cleanup because
the interface forced the use of an arena allocator. This meant that
memory was always cleaned up at program exit, so deinits/frees were not
necessary within the code.

However, this is not very idiomatic, and makes testing more difficult.
Freeing memory with an arena allocator is essentially a no-op, so it
doesn't cost much to still free temporary memory even when using one.