Reduce memory allocations
Rather than duplicate the filename for every entry in the tag list, have
the Tags struct take ownership of each unique filename string. The
pointer to each filename is stored in the 'visited' hashmap, where it
can be freed when the Tags struct is deinited.

Also use a stack allocated buffer for some filepath operations, when
possible.