Require explicit allocator in Tags.write
This decouples the allocator used to write the tags from the allocator
used by the Tags object itself. The output of tags.write must be freed
by the caller, and it is more idiomatic to make it explicit *which*
allocator is doing the allocating to make sure the correct allocator is
used to free.