Be explicit that removeDuplicates operates on the array in-place
Returning a new EntryList made it possible to assign the return value to
a different variable than the input value (`orig`), even though the
semantics of this function are that the return value should replace
`orig` (because `orig` is invalidated when this function is called).
Tighten this contract by removing the return value and forcing `orig` to
be updated within the function body directly.